% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_type_df}
\alias{cell_type_df}
\title{cell_type_df}
\format{
A list of gates to use. Gates consist of x and y coordinates.
}
\usage{
data(cell_type_df)
}
\description{
Stefano's cell type df
}
\keyword{datasets}
