% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{de_gate_gates}
\alias{de_gate_gates}
\title{de_gate_gates}
\format{
A list of gates to use. Gates consist of x and y coordinates.
}
\usage{
data(de_gate_gates)
}
\description{
Stefano's gate list
}
\keyword{datasets}
