% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{pbmc_small}
\format{
An object of class \code{SingleCellExperiment} with 230 rows and 80 columns.
}
\usage{
data(pbmc_small)
}
\description{
pbmc_small
}
\keyword{datasets}
